Make repeat count non-zero to encode the existing invariant into the typesystem

+++ compiler/dd-lir/src/model.rs
@@ -33,7 +33,7 @@ pub struct BlockMethod {
 pub enum Repeat {
     None,
     Count {
-        count: u64,
+        count: u32,
         stride: i128,
     },
     Enum {
